LayoutTests/svg/W3C-SVG-1.1/interact-order-03-b-expected.txt

KCanvasRegistry: empty
KCanvasContainer at (0,0) size 479x359
  KCanvasContainer at (9,54) size 461x231 [fill={[type=SOLID] [color=#000000]}]
    KCanvasContainer at (9,54) size 461x231 [fill={[type=SOLID] [color=#000000]}]
      KCanvasItem {rect} at (9,54) size 461x231 [stroke={[type=SOLID] [color=#000000]}] [fill={[type=SOLID] [color=#EEEEEE]}] [data="M10.00,55.00L470.00,55.00L470.00,285.00L10.00,285.00"]
    KCanvasContainer at (0,0) size 0x0 [fill={[type=SOLID] [color=#000000]}]
      KCanvasContainer at (0,0) size 0x0 [fill={[type=SOLID] [color=#000000]}]
        KCanvasContainer at (0,0) size 0x0 [fill={[type=SOLID] [color=#000000]}]
  KCanvasItem {rect} at (0,0) size 479x359 [stroke={[type=SOLID] [color=#000000]}] [data="M1.00,1.00L479.00,1.00L479.00,359.00L1.00,359.00"]
